// src/enum.ts
export enum ReportTab {
    RECEIVE = 'receive',
    SALE_BILL = 'saleBill',
    SALE_DELIVERY = 'saleDelivery',
    SALE_INVOICE = 'saleInvoice',
    SALE_DEBT = 'saleDebt',
    SALE_DEBT_SUMMARY='saleDebtSummary',
    SALE_TAX='saleTax',
    LOAN_VENDOR = 'loanVendor',
    LOAN_CUSTOMER = 'loanCustomer',
    STOCK_REMAIN = 'stockRemain',
    STOCK_MOVEMENT = 'stockMovement',
    TAX='tax'
}
export enum ReportType {
    RECEIVE = 'receive',
    SALE = 'sale',
    LOAN = 'loan',
    STOCK = 'stock',
    TAX = 'tax,'
}
export enum Month {
    JANUARY = 'มกราคม', 
    FEBRUARY = 'กุมภาพันธ์', 
    MARCH = 'มีนาคม',    
    APRIL = 'เมษายน',    
    MAY = 'พฤษภาคม',   
    JUNE = 'มิถุนายน',    
    JULY = 'กรกฎาคม',    
    AUGUST = 'สิงหาคม',   
    SEPTEMBER = 'กันยายน',  
    OCTOBER = 'ตุลาคม',    
    NOVEMBER = 'พฤศจิกายน',  
    DECEMBER = 'ธันวาคม',  
  }